+++ src/player/noteplayer.cpp
@@ -28,7 +28,7 @@ NotePlayer::~NotePlayer() {
 
 
 float NotePlayer::getSample(float srate) {
-    float sum = 0;
+    double sum = 0;
     int i;
     for(auto it = m_notes.begin(); it != m_notes.end(); std::advance(it, 1)) {
         if(it->isComplete()) {
@@ -37,10 +37,10 @@ float NotePlayer::getSample(float srate) {
             std::advance(it, -1);
             continue;
         }
-        float t = it->getTime();
+        double t = it->getTime();
         Generator& gen = *m_gen;
 
-        float sample = gen(std::distance(m_notes.begin(), it), it->getDelta(srate), t);
+        double sample = gen(std::distance(m_notes.begin(), it), it->getDelta(srate), t);
         sum += (*it).getAmplitude() * sample;
         
     }
